Make some fixes from previous PR comments
