+++ static/app/views/performance/transactionSummary/transactionOverview/relatedIssues.tsx
@@ -14,9 +14,10 @@ import {t, tct} from 'app/locale';
 import space from 'app/styles/space';
 import {OrganizationSummary} from 'app/types';
 import {trackAnalyticsEvent} from 'app/utils/analytics';
-import {TRACING_FIELDS} from 'app/utils/discover/fields';
 import {decodeScalar} from 'app/utils/queryString';
 import {MutableSearch} from 'app/utils/tokenizeSearch';
+
+import {removeTracingKeysFromSearch} from '../../utils';
 
 type Props = {
   organization: OrganizationSummary;
@@ -26,14 +27,6 @@ type Props = {
   start?: string;
   end?: string;
 };
-
-const EXCLUDE_TAG_KEYS = new Set([
-  // event type can be "transaction" but we're searching for issues
-  'event.type',
-  // the project is already determined by the transaction,
-  // and issue search does not support the project filter
-  'project',
-]);
 
 class RelatedIssues extends Component<Props> {
   getIssuesEndpoint() {
@@ -48,20 +41,7 @@ class RelatedIssues extends Component<Props> {
       ...pick(location.query, [...Object.values(URL_PARAM), 'cursor']),
     };
     const currentFilter = new MutableSearch(decodeScalar(location.query.query, ''));
-    currentFilter.getFilterKeys().forEach(tagKey => {
-      const searchKey = tagKey.startsWith('!') ? tagKey.substr(1) : tagKey;
-      // Remove aggregates and transaction event fields
-      if (
-        // aggregates
-        searchKey.match(/\w+\(.*\)/) ||
-        // transaction event fields
-        TRACING_FIELDS.includes(searchKey) ||
-        // tags that we don't want to pass to pass to issue search
-        EXCLUDE_TAG_KEYS.has(searchKey)
-      ) {
-        currentFilter.removeFilter(tagKey);
-      }
-    });
+    removeTracingKeysFromSearch(currentFilter);
     currentFilter
       .addFreeText('is:unresolved')
       .setFilterValues('transaction', [transaction]);
